§1.3 (GSOP Plasticity)
CUDA C++ Structs: The VariantParameters and Constant Memory structures are now in exact byte alignment with Rust (fixed garbage reading). The global LUT has been removed in favor of a per-type inertia curve.

Errors in CUDA: Hardcoded slot limit removed (80 \-> p.ltm_slot_count). Bit shift inertia fixed (>> 12 \-> >> 11).

Rust types: prune_threshold is now universally signed i16 (previously it was u16 in the config and implicitly cast). NeuronType now has #[derive(Default)].

GSOP math: Wrote a CPU emulator and 7 tests genesis-core/src/test_gsop_math.rs. There were difficulties with bitwise shifting of signed numbers (Rust implements arithmetic shifting, which resulted in a difference of 1 unit on depressions: -37 vs -38), but now the tests copy the logic of physics.cu as a reference. All 7 tests PASS, nvcc build also PASS.

Documents: task.md and walkthrough.md have been updated.

+++ genesis-runtime/src/main.rs
@@ -134,7 +134,7 @@ async fn main() -> Result<()> {
 
         // 4.3 Setup ZoneRuntime
         let const_mem = ZoneRuntime::build_constant_memory(&blueprints);
-        let prune_threshold = blueprints.neuron_types.first().map(|n| n.prune_threshold as i16).unwrap_or(15);
+        let prune_threshold = blueprints.neuron_types.first().map(|n| n.prune_threshold).unwrap_or(15);
         let v_seg = (sim_config.simulation.signal_speed_um_tick / sim_config.simulation.voxel_size_um) as u32;
         let master_seed = genesis_core::seed::MasterSeed::from_str(&sim_config.simulation.master_seed);
